+++ packages/core/src/llm/model/router/openai-compatible/index.ts
@@ -1,0 +1,512 @@
+import type {
+  LanguageModelV2CallOptions,
+  LanguageModelV2FinishReason,
+  LanguageModelV2StreamPart,
+  LanguageModelV2CallWarning,
+} from '@ai-sdk/provider-v5';
+import type { ResolvedModelConfig } from '../../gateway-resolver.js';
+import type { MastraModelGateway } from '../../gateways/base.js';
+import { getProviderConfig } from '../../provider-registry.generated.js';
+import type { ModelRouterLanguageModelBase, ResolvedRequestConfig } from '../base.js';
+
+// TODO: get these types from openai
+interface OpenAIStreamChunk {
+  id?: string;
+  object?: string;
+  created?: number;
+  model?: string;
+  choices?: Array<{
+    index: number;
+    delta?: {
+      role?: string;
+      content?: string | null;
+      tool_calls?: Array<{
+        index: number;
+        id?: string;
+        type?: string;
+        function?: {
+          name?: string;
+          arguments?: string;
+        };
+      }>;
+    };
+    finish_reason?: string | null;
+  }>;
+  usage?: {
+    prompt_tokens?: number;
+    completion_tokens?: number;
+    total_tokens?: number;
+  };
+}
+
+export class OpenAICompatibleModel implements ModelRouterLanguageModelBase {
+  constructor(
+    public provider: string,
+    public gateway: MastraModelGateway,
+  ) {}
+
+  public async resolveRequestConfig(
+    config: ResolvedModelConfig,
+    apiKey: string,
+  ): Promise<{ url: string; headers: Record<string, string>; modelId: string }> {
+    // If they provide a url as a string or in the config object, we shouldn't use our gateway classes
+    const shouldUseGateway = !config.url;
+
+    if (shouldUseGateway) {
+      // Use gateway resolution - always get fresh values
+      const { url, headers, resolvedModelId } = config;
+
+      if (url === false) {
+        throw new Error(`No gateway can handle model: ${config.fullModelId}`);
+      }
+
+      // Merge headers with custom headers
+      const finalHeaders = {
+        'Content-Type': 'application/json',
+        ...headers,
+        ...config.headers,
+      };
+
+      return { url, headers: finalHeaders, modelId: resolvedModelId };
+    } else {
+      // Use static configuration
+      if (!config.url) {
+        throw new Error('URL is required for OpenAI-compatible model');
+      }
+
+      const headers: Record<string, string> = {
+        'Content-Type': 'application/json',
+        ...config.headers,
+      };
+
+      // Add auth header if API key is available
+      if (apiKey) {
+        // Check if we need a special header format
+        const providerConfig = this.provider !== 'openai-compatible' ? getProviderConfig(this.provider) : undefined;
+        if (providerConfig?.apiKeyHeader === 'x-api-key') {
+          headers['x-api-key'] = apiKey;
+        } else {
+          headers['Authorization'] = `Bearer ${apiKey}`;
+        }
+      }
+
+      return { url: config.url, headers, modelId: config.resolvedModelId };
+    }
+  }
+
+  private convertMessagesToOpenAI(messages: LanguageModelV2CallOptions['prompt']): any[] {
+    return messages
+      .map(msg => {
+        if (msg.role === 'system') {
+          return {
+            role: 'system',
+            content: msg.content,
+          };
+        }
+
+        if (msg.role === 'user') {
+          // Handle content parts
+          const contentParts = msg.content
+            .map(part => {
+              if (part.type === 'text') {
+                return { type: 'text', text: part.text };
+              }
+              // Note: v2 uses 'file' type with image property
+              if (part.type === 'file') {
+                return {
+                  type: 'image_url',
+                  image_url: { url: part.data },
+                };
+              }
+              return null;
+            })
+            .filter(Boolean);
+
+          // If only text parts, flatten to string
+          if (contentParts.every(p => p?.type === 'text')) {
+            return {
+              role: 'user',
+              content: contentParts.map(p => p?.text || '').join(''),
+            };
+          }
+
+          return {
+            role: 'user',
+            content: contentParts,
+          };
+        }
+
+        if (msg.role === 'assistant') {
+          const textContent = msg.content
+            .filter(part => part.type === 'text')
+            .map(part => part.text)
+            .join('');
+
+          const toolCalls = msg.content
+            .filter(part => part.type === 'tool-call')
+            .map(part => ({
+              id: part.toolCallId,
+              type: 'function',
+              function: {
+                name: part.toolName,
+                arguments: JSON.stringify(part.input || {}),
+              },
+            }));
+
+          return {
+            role: 'assistant',
+            content: textContent || null,
+            ...(toolCalls.length > 0 && { tool_calls: toolCalls }),
+          };
+        }
+
+        if (msg.role === 'tool') {
+          return msg.content.map(toolResponse => ({
+            role: 'tool',
+            tool_call_id: toolResponse.toolCallId,
+            content: JSON.stringify(toolResponse.output),
+          }));
+        }
+
+        return msg;
+      })
+      .flat();
+  }
+
+  public convertTools(tools: LanguageModelV2CallOptions['tools']): any[] | undefined {
+    if (!tools || Object.keys(tools).length === 0) return undefined;
+
+    return Object.entries(tools).map(([name, tool]) => {
+      if (tool.type === 'function') {
+        return {
+          type: 'function',
+          function: {
+            name: tool.name,
+            description: tool.description,
+            parameters: tool.inputSchema || {},
+          },
+        };
+      }
+      // For provider-defined tools, use minimal definition
+      return {
+        type: 'function',
+        function: {
+          name,
+          description: `Provider tool: ${name}`,
+          parameters: {},
+        },
+      };
+    });
+  }
+
+  public mapFinishReason(reason: string | null): LanguageModelV2FinishReason {
+    switch (reason) {
+      case 'stop':
+        return 'stop';
+      case 'length':
+      case 'max_tokens':
+        return 'length';
+      case 'tool_calls':
+      case 'function_call':
+        return 'tool-calls';
+      case 'content_filter':
+        return 'content-filter';
+      default:
+        return 'unknown';
+    }
+  }
+
+  async doStream(
+    options: LanguageModelV2CallOptions,
+    config: ResolvedRequestConfig,
+  ): Promise<{
+    stream: ReadableStream<LanguageModelV2StreamPart>;
+    request?: { body: string };
+    response?: { headers: Record<string, string> };
+    warnings: LanguageModelV2CallWarning[];
+  }> {
+    try {
+      const { prompt, tools, toolChoice, providerOptions } = options;
+      const { url, headers, modelId: resolvedModelId } = config;
+
+      // TODO: real body type, not any
+      const body: any = {
+        messages: this.convertMessagesToOpenAI(prompt),
+        model: resolvedModelId,
+        stream: true,
+        ...providerOptions,
+      };
+
+      const openAITools = this.convertTools(tools);
+      if (openAITools) {
+        body.tools = openAITools;
+        if (toolChoice) {
+          body.tool_choice =
+            toolChoice.type === 'none'
+              ? 'none'
+              : toolChoice.type === 'required'
+                ? 'required'
+                : toolChoice.type === 'auto'
+                  ? 'auto'
+                  : toolChoice.type === 'tool'
+                    ? { type: 'function', function: { name: toolChoice.toolName } }
+                    : 'auto';
+        }
+      }
+
+      // Handle structured output
+      if (options.responseFormat?.type === 'json') {
+        body.response_format = {
+          type: 'json_schema',
+          json_schema: {
+            name: 'response',
+            strict: true,
+            schema: options.responseFormat.schema,
+          },
+        };
+      }
+
+      const fetchArgs = {
+        method: 'POST',
+        headers,
+        body: JSON.stringify(body),
+        signal: options.abortSignal,
+      };
+      const response = await fetch(url, fetchArgs);
+
+      if (!response.ok) {
+        const error = await response.text();
+
+        // Check for authentication errors
+        if (response.status === 401 || response.status === 403) {
+          const providerConfig = getProviderConfig(this.provider);
+          if (providerConfig?.apiKeyEnvVar) {
+            throw new Error(
+              `Authentication failed for provider "${this.provider}". Please ensure the ${providerConfig.apiKeyEnvVar} environment variable is set with a valid API key.`,
+            );
+          }
+        }
+
+        throw new Error(`OpenAI-compatible API error: ${response.status} - ${error}`);
+      }
+
+      const reader = response.body?.getReader();
+      if (!reader) {
+        throw new Error('Response body is not readable');
+      }
+
+      const decoder = new TextDecoder();
+      let buffer = '';
+      let sentStart = false;
+      const toolCallBuffers = new Map<number, { id: string; name: string; args: string; sent?: boolean }>();
+      const mapFinishReason = this.mapFinishReason.bind(this);
+
+      let isActiveText = false;
+
+      const stream = new ReadableStream<LanguageModelV2StreamPart>({
+        async start(controller) {
+          try {
+            // Send stream-start
+            controller.enqueue({
+              type: 'stream-start',
+              warnings: [],
+            });
+
+            while (true) {
+              const { done, value } = await reader.read();
+
+              if (done) {
+                // Parse and send any buffered tool calls that haven't been sent yet
+                for (const [_, toolCall] of toolCallBuffers) {
+                  if (!toolCall.sent && toolCall.id && toolCall.name && toolCall.args) {
+                    controller.enqueue({
+                      type: 'tool-call',
+                      toolCallId: toolCall.id,
+                      toolName: toolCall.name,
+                      input: toolCall.args || '{}',
+                    });
+                  }
+                }
+
+                controller.close();
+                break;
+              }
+
+              buffer += decoder.decode(value, { stream: true });
+              const lines = buffer.split('\n');
+              buffer = lines.pop() || '';
+
+              for (const line of lines) {
+                if (line.trim() === '' || line.trim() === 'data: [DONE]') {
+                  continue;
+                }
+
+                if (line.startsWith('data: ')) {
+                  try {
+                    const data: OpenAIStreamChunk = JSON.parse(line.slice(6));
+
+                    // Send response metadata from first chunk
+                    if (!sentStart && data.id) {
+                      controller.enqueue({
+                        type: 'response-metadata',
+                        id: data.id,
+                        modelId: data.model || config.modelId,
+                        timestamp: new Date(data.created ? data.created * 1000 : Date.now()),
+                      });
+                      sentStart = true;
+                    }
+
+                    const choice = data.choices?.[0];
+                    if (!choice) continue;
+
+                    // Handle text delta
+                    if (choice.delta?.content) {
+                      if (!isActiveText) {
+                        controller.enqueue({ type: 'text-start', id: 'text-1' });
+                        isActiveText = true;
+                      }
+
+                      controller.enqueue({
+                        type: 'text-delta',
+                        id: 'text-1',
+                        delta: choice.delta.content,
+                      });
+                    } else if (isActiveText) {
+                      controller.enqueue({ type: 'text-end', id: 'text-1' });
+                      isActiveText = false;
+                    }
+
+                    // Handle tool call deltas
+                    if (choice.delta?.tool_calls) {
+                      for (const toolCall of choice.delta.tool_calls) {
+                        const index = toolCall.index;
+
+                        if (!toolCallBuffers.has(index)) {
+                          // Send tool-input-start when we first see a tool call
+                          if (toolCall.id && toolCall.function?.name) {
+                            controller.enqueue({
+                              type: 'tool-input-start',
+                              id: toolCall.id,
+                              toolName: toolCall.function.name,
+                            });
+                          }
+
+                          toolCallBuffers.set(index, {
+                            id: toolCall.id || '',
+                            name: toolCall.function?.name || '',
+                            args: '',
+                          });
+                        }
+
+                        const buffer = toolCallBuffers.get(index)!;
+
+                        if (toolCall.id) {
+                          buffer.id = toolCall.id;
+                        }
+
+                        if (toolCall.function?.name) {
+                          buffer.name = toolCall.function.name;
+                        }
+
+                        if (toolCall.function?.arguments) {
+                          buffer.args += toolCall.function.arguments;
+                          controller.enqueue({
+                            type: 'tool-input-delta',
+                            id: buffer.id,
+                            delta: toolCall.function.arguments,
+                          });
+
+                          // Check if tool call is complete (parsable JSON)
+                          try {
+                            JSON.parse(buffer.args);
+                            if (buffer.id && buffer.name) {
+                              controller.enqueue({
+                                type: 'tool-input-end',
+                                id: buffer.id,
+                              });
+
+                              controller.enqueue({
+                                type: 'tool-call',
+                                toolCallId: buffer.id,
+                                toolName: buffer.name,
+                                input: buffer.args,
+                              });
+
+                              toolCallBuffers.set(index, {
+                                id: buffer.id,
+                                name: buffer.name,
+                                args: buffer.args,
+                                sent: true,
+                              });
+                            }
+                          } catch {
+                            // Not complete JSON yet, continue buffering
+                          }
+                        }
+                      }
+                    }
+
+                    // Handle finish
+                    if (choice.finish_reason) {
+                      // Don't send tool calls again - they've already been sent when complete
+                      toolCallBuffers.clear();
+
+                      controller.enqueue({
+                        type: 'finish',
+                        finishReason: mapFinishReason(choice.finish_reason),
+                        usage: data.usage
+                          ? {
+                              inputTokens: data.usage.prompt_tokens || 0,
+                              outputTokens: data.usage.completion_tokens || 0,
+                              totalTokens: data.usage.total_tokens || 0,
+                            }
+                          : {
+                              inputTokens: 0,
+                              outputTokens: 0,
+                              totalTokens: 0,
+                            },
+                      });
+                    }
+                  } catch (e) {
+                    console.error('Error parsing SSE data:', e);
+                  }
+                }
+              }
+            }
+          } catch (error) {
+            return {
+              stream: new ReadableStream({
+                start(controller) {
+                  controller.enqueue({
+                    type: 'error',
+                    error: error instanceof Error ? error.message : String(error),
+                  } as LanguageModelV2StreamPart);
+                },
+              }),
+              warnings: [],
+            };
+          }
+        },
+      });
+
+      return {
+        stream,
+        request: { body: JSON.stringify(body) },
+        response: { headers: Object.fromEntries(response.headers.entries()) },
+        warnings: [],
+      };
+    } catch (error) {
+      return {
+        stream: new ReadableStream({
+          start(controller) {
+            controller.enqueue({
+              type: 'error',
+              error: error instanceof Error ? error.message : String(error),
+            } as LanguageModelV2StreamPart);
+          },
+        }),
+        warnings: [],
+      };
+    }
+  }
+}
